refactor: move AdoptTree to components/trees and remove features/adoption
