+++ frontend/src/components/trees/WaterColumn.svelte
