+++ frontend/src/components/ui/index.ts
@@ -1,6 +1,6 @@
 export { default as Accordion } from './Accordion.svelte';
 export { default as Button } from './Button.svelte';
-export { default as Heading } from './Heading.svelte';
-export { default as InputField } from './InputField.svelte';
-export { default as EmailField } from './EmailField.svelte';
-export { default as PasswordField } from './PasswordField.svelte';
+export { default as Heading } from './Heading.svelte'; 
+export { default as InputField } from './input/InputField.svelte';
+export { default as EmailField } from './input/EmailField.svelte';
+export { default as PasswordField } from './input/PasswordField.svelte';
