+++ frontend/src/routes/trees/[treeId]/+page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import { Accordion } from '$components/ui';
-	import WaterColumn from '$components/WaterColumn.svelte';
+	import WaterColumn from '$components/trees/WaterColumn.svelte';
 	import Chat from '$components/chat/Chat.svelte';
 	import { DialogPanel } from '$components/overlay';
 
-	import AdoptTree from '../../../features/adoption/AdoptTree.svelte';
+	import AdoptTree from '$components/trees/AdoptTree.svelte';
 	import { supabase } from '../../../supabase';
 	import { onMount } from 'svelte';
 	import { page } from '$app/stores';
